pyproject.toml
[project]
name = "segb"
version = "0.2.0"
description = "Semantic Ethical Glass Box (SEGB) for AMOR project"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi[standard]==0.115.12",
    "exceptiongroup>=1.2.2; python_version < '3.11'",
    "PyYAML==6.0.2",
    "uvicorn[standard]==0.34.0",
    "pyjwt==2.10.1",
    "rdflib==7.1.3",
    "requests==2.31.0",
    "python-multipart>=0.0.9"
]

[tool.ruff]
line-length = 120
target-version = "py310"
exclude = [".venv"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "C4", "SIM"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
check_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
ignore_missing_imports = true
exclude = "^(\\.venv|examples/notebooks/)"

[tool.pytest.ini_options]
testpaths = [
    "apps/backend/tests",
    "packages/semantic_log_generator/tests",
    "examples/simulations/tests",
    "tests",
]
python_files = ["test_*.py"]
addopts = "-ra"

[tool.coverage.run]
branch = true
source = ["apps", "packages", "examples"]

[tool.coverage.report]
show_missing = true
skip_covered = true
